Change AuroraController to an interface
